layout event moved to BaseRoomSession
