+++ packages/js/src/utils/interfaces/callFabric.ts
@@ -1,14 +1,4 @@
-import { VideoLayoutChangedEventParams, VideoPosition } from '@signalwire/core'
-
 export interface CallFabricRoomSessionContract {
-  /**
-   * The layout returned from the `layout.changed` event based on the current room layout
-   */
-  currentLayout: VideoLayoutChangedEventParams['layout']
-  /**
-   * The current position of the member returned from the `layout.changed` event
-   */
-  currentPosition: VideoPosition
   /**
    * Starts the call via the WebRTC connection
    *
